+++ project/editor/core/face-library/builtin/index.js
@@ -14,5 +14,6 @@ import { BROW_SETS } from './brows.js';
 import { NOSES } from './noses.js';
 import { MOUTHS } from './mouths.js';
 import { EAR_SETS } from './ears.js';
+import { HAIR_STYLES } from './hair.js';
 
-export const BUILTIN_FACE_PARTS = Object.freeze([...HEADS, ...EYE_SETS, ...BROW_SETS, NOSE_DOT, ...NOSES, MOUTH_SIMPLE, MOUTH_WIDE, ...MOUTHS, ...EAR_SETS]);
+export const BUILTIN_FACE_PARTS = Object.freeze([...HEADS, ...EYE_SETS, ...BROW_SETS, NOSE_DOT, ...NOSES, MOUTH_SIMPLE, MOUTH_WIDE, ...MOUTHS, ...EAR_SETS, ...HAIR_STYLES]);